Move local_ledger migration to blockprocessor.
