+++ processor/blockprocessor/block_processor.go
@@ -2,7 +2,6 @@ package blockprocessor
 
 import (
 	"fmt"
-
 	log "github.com/sirupsen/logrus"
 
 	"github.com/algorand/go-algorand/config"
@@ -39,30 +38,27 @@ func MakeProcessorWithLedger(l *ledger.Ledger, handler func(block *ledgercore.Va
 
 // MakeProcessorWithLedgerInit creates a block processor and initializes the ledger.
 func MakeProcessorWithLedgerInit(logger *log.Logger, catchpoint string, genesis *bookkeeping.Genesis, nextDBRound uint64, opts idb.IndexerDbOptions, handler func(block *ledgercore.ValidatedBlock) error) (processor.Processor, error) {
-	/*
-		// TODO: Uncomment this once migrations/local_ledger is moved into this package to avoid circular dependency.
-			if nextDBRound > 0 {
-				if catchpoint != "" {
-					round, _, err := ledgercore.ParseCatchpointLabel(catchpoint)
-					if err != nil {
-						return &blockProcessor{}, fmt.Errorf("MakeProcessorWithCatchup() label err: %w", err)
-					}
-					if uint64(round) >= nextDBRound {
-						logger.Warnf("round for given catchpoint is ahead of db round. skip fast catchup")
-					} else {
-						err = localledger.RunMigrationFastCatchup(logging.NewLogger(), catchpoint, opts.IndexerDatadir, *genesis)
-						if err != nil {
-							return &blockProcessor{}, fmt.Errorf("MakeProcessorWithCatchup() fast catchup err: %w", err)
-						}
-					}
-
-				}
-				err := localledger.RunMigrationSimple(logger, nextDBRound-1, &opts)
+	if nextDBRound > 0 {
+		if catchpoint != "" {
+			round, _, err := ledgercore.ParseCatchpointLabel(catchpoint)
+			if err != nil {
+				return &blockProcessor{}, fmt.Errorf("MakeProcessorWithCatchup() label err: %w", err)
+			}
+			if uint64(round) >= nextDBRound {
+				logger.Warnf("round for given catchpoint is ahead of db round. skip fast catchup")
+			} else {
+				err = InitializeLedgerFastCatchup(logger, catchpoint, opts.IndexerDatadir, *genesis)
 				if err != nil {
-					return &blockProcessor{}, fmt.Errorf("MakeProcessorWithCatchup() slow catchup err: %w", err)
+					return &blockProcessor{}, fmt.Errorf("MakeProcessorWithCatchup() fast catchup err: %w", err)
 				}
 			}
-	*/
+
+		}
+		err := InitializeLedgerSimple(logger, nextDBRound-1, &opts)
+		if err != nil {
+			return &blockProcessor{}, fmt.Errorf("MakeProcessorWithCatchup() slow catchup err: %w", err)
+		}
+	}
 	return MakeProcessor(logger, genesis, nextDBRound, opts.AlgodDataDir, handler)
 }
